migrate examples to EntryPoint v0.9 defaults
Default the Safe examples to SafeMultiChainSigAccountV1 (Safe Unified
Account) and the Simple7702 examples to Simple7702AccountV09, both on
EntryPoint v0.9. The Unified Account works single-chain and across chains
with one class.

- Safe examples: swap SafeAccountV0_3_0 -> SafeMultiChainSigAccountV1;
  fold batch-transactions-1-5-0 into batch-transactions
- eip-712-signing: drive getUserOperationEip712Data manually and format
  with isMultiChainSignature (Unified Account validates single ops via
  the multi-chain scheme)
- sponsor-gas: use Erc7677Paymaster
- eip-7702/simple-account: default to Simple7702AccountV09; fold the
  dedicated v0.9 duplicates (03, 06) into the base examples; add
  03-parallel-signing.ts to showcase the two-phase commit/finalize flow;
  renumber so files stay contiguous (01-07)
- docs: AGENTS.md account-types + goal table, README default-account note,
  eip-7702 and signer READMEs
