+++ eip-7702/README.md
@@ -6,14 +6,14 @@ These examples demonstrate how to upgrade an EOA to a Smart Account with EIP-770
 
 ```text
 eip-7702/
-├── simple-account/       # Simple7702Account examples
-│   ├── 01-upgrade-eoa.ts            # Upgrade EOA + batch mint NFTs (sponsored gas)
+├── simple-account/       # Simple7702AccountV09 examples (EntryPoint v0.9)
+│   ├── 01-upgrade-eoa.ts            # Upgrade EOA + mint NFTs (sponsored gas)
 │   ├── 02-upgrade-eoa-erc20-gas.ts  # Upgrade EOA with ERC-20 token gas payment
-│   ├── 03-upgrade-eoa-ep-v09.ts     # Upgrade EOA using EntryPoint v0.9
+│   ├── 03-parallel-signing.ts       # Two-phase paymaster signing (commit/finalize) latency optimization
 │   ├── 04-revoke-delegation.ts      # Revoke EIP-7702 delegation
-│   ├── 05-external-signer.ts        # Upgrade EOA with the v0.3.2 ExternalSigner API
-│   ├── 06-external-signer-v09.ts    # Same, EntryPoint v0.9 (two-phase paymaster)
-│   └── 07-typed-data-builder.ts     # Drive signTypedData yourself via getUserOperationEip712Data
+│   ├── 05-external-signer.ts        # Upgrade EOA with the ExternalSigner API
+│   ├── 06-typed-data-builder.ts     # Drive signTypedData yourself via getUserOperationEip712Data
+│   └── 07-migrate-v08-to-v09.ts     # Migrate an existing v0.8 delegation to v0.9
 └── calibur-account/      # Calibur7702Account examples (passkeys, key management)
     ├── 01-upgrade-eoa.ts
     ├── 02-passkeys.ts
@@ -81,10 +81,9 @@ userOp.signature = await smartAccount.signUserOperationWithSigner(
 )
 ```
 
-See `simple-account/05-external-signer.ts` for the end-to-end Simple7702 flow,
-`simple-account/06-external-signer-v09.ts` for the EntryPoint v0.9 variant, and
-`calibur-account/04-external-signer.ts` for Calibur. The `signer/` folder at the
-repo root contains one self-contained example per adapter.
+See `simple-account/05-external-signer.ts` for the end-to-end Simple7702 flow
+and `calibur-account/04-external-signer.ts` for Calibur. The `signer/` folder at
+the repo root contains one self-contained example per adapter.
 
 Note: the **delegation authorization** (`createAndSignEip7702DelegationAuthorization`)
 still takes the raw private key. It is a separate signature from the UserOperation
