+++ eip-7702/simple-account/01-upgrade-eoa.ts
@@ -1,104 +1,85 @@
 import { loadEnv, getOrCreateOwner } from '../../utils/env'
 import {
-    Simple7702Account,
+    Simple7702AccountV09 as Simple7702Account,
+    MetaTransaction,
+    UserOperationV9,
     getFunctionSelector,
     createCallData,
     createAndSignEip7702DelegationAuthorization,
     CandidePaymaster,
 } from "abstractionkit"
 
-// Upgrades an EOA to a Simple7702 smart account with gas sponsorship,
-// then batch-mints two NFTs in a single UserOperation.
+// Upgrades an EOA to a Simple7702 smart account (EntryPoint v0.9) with gas
+// sponsorship, then mints two NFTs across two separate UserOperations (one
+// after the other) so we exercise both the "first userOp (with 7702 auth)"
+// path and the "subsequent userOp (already delegated)" path.
 
 async function main(): Promise<void> {
     const { chainId, bundlerUrl, nodeUrl, paymasterUrl, sponsorshipPolicyId } = loadEnv()
-    const { publicAddress: eoaDelegatorPublicAddress, privateKey: eoaDelegatorPrivateKey } = getOrCreateOwner()
+    const { publicAddress: eoaAddress, privateKey } = getOrCreateOwner()
 
-    // ──────────────────────────────────────────────────────────────────────
-    // Step 1: Initialize smart account and build transactions
-    // ──────────────────────────────────────────────────────────────────────
-    // We'll batch-mint two NFTs in a single UserOperation to demonstrate
-    // the smart account's batching capability.
-    const smartAccount = new Simple7702Account(eoaDelegatorPublicAddress)
+    const smartAccount = new Simple7702Account(eoaAddress)
+    const paymaster = new CandidePaymaster(paymasterUrl)
 
     const nftContractAddress = "0x9a7af758aE5d7B6aAE84fe4C5Ba67c041dFE5336"
-    const mintFunctionSelector = getFunctionSelector('mint(address)')
     const mintCallData = createCallData(
-        mintFunctionSelector,
+        getFunctionSelector('mint(address)'),
         ["address"],
-        [eoaDelegatorPublicAddress],
+        [eoaAddress],
     )
+    const mintTx: MetaTransaction = { to: nftContractAddress, value: 0n, data: mintCallData }
 
-    const mintNft1 = { to: nftContractAddress, value: 0n, data: mintCallData }
-    const mintNft2 = { to: nftContractAddress, value: 0n, data: mintCallData }
+    async function sendOneMint(label: string): Promise<void> {
+        console.log(`\n========== ${label} ==========`)
 
-    // ──────────────────────────────────────────────────────────────────────
-    // Step 2: Create UserOperation with EIP-7702 authorization
-    // ──────────────────────────────────────────────────────────────────────
-    let userOperation = await smartAccount.createUserOperation(
-        [mintNft1, mintNft2],
-        nodeUrl,
-        bundlerUrl,
-        {
-            eip7702Auth: {
-                chainId: chainId,
-            }
-        }
-    )
-
-    // ──────────────────────────────────────────────────────────────────────
-    // Step 3: Sign EIP-7702 delegation authorization
-    // ──────────────────────────────────────────────────────────────────────
-    // If the EOA is already delegated, eip7702Auth is null — skip signing.
-    if (userOperation.eip7702Auth) {
-        userOperation.eip7702Auth = createAndSignEip7702DelegationAuthorization(
-            BigInt(userOperation.eip7702Auth.chainId),
-            userOperation.eip7702Auth.address,
-            BigInt(userOperation.eip7702Auth.nonce),
-            eoaDelegatorPrivateKey
+        let userOperation: UserOperationV9 = await smartAccount.createUserOperation(
+            [mintTx],
+            nodeUrl,
+            bundlerUrl,
+            { eip7702Auth: { chainId } },
         )
+
+        if (userOperation.eip7702Auth) {
+            console.log("EOA not yet delegated — signing EIP-7702 authorization")
+            userOperation.eip7702Auth = createAndSignEip7702DelegationAuthorization(
+                BigInt(userOperation.eip7702Auth.chainId),
+                userOperation.eip7702Auth.address,
+                BigInt(userOperation.eip7702Auth.nonce),
+                privateKey,
+            )
+        } else {
+            console.log("EOA already delegated — skipping EIP-7702 authorization")
+        }
+
+        const { userOperation: sponsored } = await paymaster.createSponsorPaymasterUserOperation(
+            smartAccount, userOperation, bundlerUrl, sponsorshipPolicyId,
+        )
+        userOperation = sponsored
+
+        userOperation.signature = smartAccount.signUserOperation(
+            userOperation, privateKey, chainId,
+        )
+
+        const response = await smartAccount.sendUserOperation(userOperation, bundlerUrl)
+        console.log(`UserOp sent (hash: ${response.userOperationHash}). Waiting for inclusion...`)
+        const receipt = await response.included()
+
+        if (receipt == null) {
+            console.log("Receipt not found (timeout)")
+            throw new Error(`${label}: receipt timeout`)
+        }
+        if (!receipt.success) {
+            console.log("UserOperation execution failed")
+            console.log(receipt)
+            throw new Error(`${label}: userOp execution failed`)
+        }
+        console.log(`${label} success — tx ${receipt.receipt.transactionHash} actualGasUsed=${receipt.actualGasUsed}`)
     }
 
-    // ──────────────────────────────────────────────────────────────────────
-    // Step 4: Sponsor gas with paymaster
-    // ──────────────────────────────────────────────────────────────────────
-    const paymaster = new CandidePaymaster(paymasterUrl)
-    const { userOperation: paymasterUserOperation } = await paymaster.createSponsorPaymasterUserOperation(
-        smartAccount, userOperation, bundlerUrl, sponsorshipPolicyId)
-    userOperation = paymasterUserOperation
+    await sendOneMint("UserOp #1 — first mint (upgrades EOA via EIP-7702)")
+    await sendOneMint("UserOp #2 — second mint (EOA already delegated)")
 
-    // ──────────────────────────────────────────────────────────────────────
-    // Step 5: Sign UserOperation
-    // ──────────────────────────────────────────────────────────────────────
-    userOperation.signature = smartAccount.signUserOperation(
-        userOperation,
-        eoaDelegatorPrivateKey,
-        chainId,
-    )
-
-    // ──────────────────────────────────────────────────────────────────────
-    // Step 6: Send and wait for inclusion
-    // ──────────────────────────────────────────────────────────────────────
-    console.log("UserOperation:", userOperation)
-
-    const sendUserOperationResponse = await smartAccount.sendUserOperation(
-        userOperation, bundlerUrl
-    )
-
-    console.log("UserOp sent! Waiting for inclusion...")
-    console.log("UserOp hash:", sendUserOperationResponse.userOperationHash)
-
-    const receipt = await sendUserOperationResponse.included()
-
-    console.log("UserOperation receipt received.")
-    console.log(receipt)
-    if (receipt == null) {
-        console.log("Receipt not found (timeout)")
-    } else if (receipt.success) {
-        console.log("EOA upgraded to a Smart Account and minted two NFTs! Transaction hash: " + receipt.receipt.transactionHash)
-    } else {
-        console.log("UserOperation execution failed")
-    }
+    console.log("\nBoth UserOperations included on EntryPoint v0.9.")
 }
 
 main()
