+++ erc7677/sponsor-gas.ts
@@ -1,6 +1,6 @@
 import { loadEnv, getOrCreateOwner } from '../utils/env'
 import {
-    SafeAccountV0_3_0 as SafeAccount,
+    SafeMultiChainSigAccountV1 as SafeAccount,
     MetaTransaction,
     Erc7677Paymaster,
     calculateUserOperationMaxGasCost,
