+++ signer/customSigner.ts
@@ -19,7 +19,7 @@ import {
     Erc7677Paymaster,
     ExternalSigner,
     MetaTransaction,
-    SafeAccountV0_3_0 as SafeAccount,
+    SafeMultiChainSigAccountV1 as SafeAccount,
     createCallData,
     getFunctionSelector,
 } from 'abstractionkit'
